+++ packages/git/src/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { type GitFs, type GitStat, vfsToGitFs } from './fs-adapter.ts';